Fix : 코드 보여주는 방식 변경
* int 초 단위 -> float 초 단위(0.1)

+++ ChordBox_android/app/src/main/java/com/misaengfly/chordbox/database/Record.kt
@@ -25,14 +25,14 @@ data class Record(
 
 fun List<Record>.asDomainModel(): List<MusicItem> {
     return map {
-        val tempMap: MutableMap<Int, String> = mutableMapOf()
+        val tempMap: MutableMap<Float, String> = mutableMapOf()
 
         if (!it.chords.isNullOrBlank()) {
             val chordList = it.chords.split(" ")
             val timeList = it.times.split(" ")
 
             for (i in chordList.indices) {
-                tempMap[timeList[i].toInt()] = chordList[i]
+                tempMap[timeList[i].toFloat()] = chordList[i]
             }
         }
